+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 - Performance and code size optimization
+- **Added:** `GlobalTlsf` now provides a `malloc`-compatible interface.
 
 ## [0.1.1] - 2021-05-23